requirements.txt
Django==1.10.4
djangorestframework==3.5.3
olefile==0.44
Pillow==4.0.0
wheel==0.24.0
